requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg[binary]==3.2.5
httpx==0.28.1
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
whitenoise==6.9.0
Faker==37.1.0
